Task #13/requirements.txt
fastapi
uvicorn[standart]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
asyncio
dotenv